Fix handling of bare DATABRICKS_HOST environment variables. (#252)
Some environments (e.g. Workbench) set `DATABRICKS_HOST` to the bare
hostname and omit the HTTPS prefix; this breaks when we later expect it
to be an actual URL.

So this commit updates `databricks_workspace()` to handle both cases.
Similar code exists in all of the other Databricks packages and SDKs
that I am aware of, so this is nothing new; I just missed it.

Unit tests are included.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>