Avoid passing a schemaless host to the Databricks CLI (#522)
Previously we were passing the host without the schema. I think this
probably worked at some point in the past, but it does not work with
modern versions of the Databricks CLI:

    $ databricks auth token --host example.cloud.databricks.com
    Error: host must start with 'https://': example.cloud.databricks.com

This commit ensures we pass the complete workspace URL. Unit tests have
been updated to verify this as well.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>