Pick up on workspace URLs set in the Databricks config file (#525)
This commit ensures we pick up on workspace URLs set in the Databricks
config file, so users that have such a configuration no longer need to
set the `DATABRICKS_HOST` environment variable. This is a nice
quality-of-life improvement for folks on desktop using the Databricks
CLI or other tooling, and brings us to parity with `chatlas`.

As part of the process I added more generic code for parsing
the Databricks config file even when not generated by Workbench. In
theory we could use this to support even more authentication methods in
the future.

Unit tests are included.

Closes #521.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>