More cleanup of merging/as_json() logic for chat_snowflake() (#556)
This merging logic is closer to what we do for Claude and Gemini and is
far more readable and maintainable.

It also paves the way to tool calling, which uses a very different input
and output format from OpenAI and is going to require a departure
anyway.

Unit tests should cover these changes.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>