Improve error handling in `parallel_chat_structured()` (#787)
Now it will warn, instead of failing, if some results are not parseable. This change includes making `ContentJson()` lazier at parsing JSON so we can during data extraction, not object creation. Also includes some tests (and a small fix) to ensure that `NULL` can be coerced to `NA`s of the correct type.

Part of #628