Refactor cost computation (#812)
Now we compute the cost once, when we create the turn. The makes a lot of things simpler, e.g. we no longer need to track the variant in `the$tokens`, and unit tests get simpler. I also introduce `Chat$get_cost_details()` (ported from #810) to make it easier to inspect the raw details of a chat.

Includes a fix for #804 by logging as part of `Chat$add_turn()`. This breaks #743 in a new way: for now batch chat usage isn't logged at all. But I think it makes sense to fix that in another PR.

I've also added `map_tokens()` helper to hide the awkward use of `vapply()` in a few places (and to make the output more consistent)